+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Fixed
 
+- fix incorrect block location on statementish op leaving trailing semicolon in last child
 - fix put star `*` to parenthesized single element non-star `ImportFrom.names`
 - fix get slice from start of unparenthesized tuple in f-string expression on py < 3.12
 - fix get format_spec escape sequences from `rf'{...:\xFF}'` string